Refactor server into server.rs
